pyproject.toml
[project]
name = "livingwp"
version = "0.1.0"
description = "GenAI Working Group's Living Whitepaper"
readme = "README.md"
authors = [{ name = "Ming", email = "ming.cheuk@elementx.ai" }]
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.0.19",
    "python-frontmatter>=1.1.0",
]

[project.scripts]
livingwp = "livingwp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["ruff>=0.11.12"]
